Dev (#179)
* final commit before rebase 2

* reset:sft

* Rebase branch (#139)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* final commit before rebase 1

* pdev:set

* temp

* removed unnecessary files

* master

* copy branch

* remove files

* Feature alexamil (#144)

* pdev:set

* pdev:set

* feature:merge:1502314677544

* hotfix

* pdev:set

* hotfix

* integrate:set

* Feature alexamil 1502315157801 (#152)

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1504130163215

* Feature alexamil 1504130287958 (#153)

* pdev:set

* feature:merge:1504208378835

* Feature alexamil 1504208519267 (#155)

* Dev (#154)

* final commit before rebase 2

* reset:sft

* Rebase branch (#139)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* final commit before rebase 1

* pdev:set

* temp

* removed unnecessary files

* master

* copy branch

* remove files

* Feature alexamil (#144)

* pdev:set

* pdev:set

* feature:merge:1502314677544

* hotfix

* pdev:set

* hotfix

* integrate:set

* Feature alexamil 1502315157801 (#152)

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1504130163215

* Feature alexamil 1504130287958 (#153)

* pdev:set

* feature:merge:1504208378835

* pdev:set

* pdev:set

* feature:merge:1504657314635

* Feature alexamil 1504657399214 (#157)

* Dev (#156)

* final commit before rebase 2

* reset:sft

* Rebase branch (#139)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* final commit before rebase 1

* pdev:set

* temp

* removed unnecessary files

* master

* copy branch

* remove files

* Feature alexamil (#144)

* pdev:set

* pdev:set

* feature:merge:1502314677544

* hotfix

* pdev:set

* hotfix

* integrate:set

* Feature alexamil 1502315157801 (#152)

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1504130163215

* Feature alexamil 1504130287958 (#153)

* pdev:set

* feature:merge:1504208378835

* Feature alexamil 1504208519267 (#155)

* Dev (#154)

* final commit before rebase 2

* reset:sft

* Rebase branch (#139)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* final commit before rebase 1

* pdev:set

* temp

* removed unnecessary files

* master

* copy branch

* remove files

* Feature alexamil (#144)

* pdev:set

* pdev:set

* feature:merge:1502314677544

* hotfix

* pdev:set

* hotfix

* integrate:set

* Feature alexamil 1502315157801 (#152)

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1504130163215

* Feature alexamil 1504130287958 (#153)

* pdev:set

* feature:merge:1504208378835

* pdev:set

* pdev:set

* feature:merge:1504657314635

* feature:merge:1504747393615

* Feature alexamil 1504747816731 (#160)

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1504910384740

* Feature alexamil 1504910440364 (#162)

* pdev:set

* feature:merge:1504984180884

* Feature alexamil 1504984265091 (#167)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1505543217998

* Feature alexamil 1505543320764 (#171)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1505969235904

* feature:merge:1505969243600

* Feature alexamil 1505969410922 (#173)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1506321415440

* Feature alexamil 1506322142732 (#176)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1506756176623

* Feature alexamil 1506756399740 (#178)

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* pdev:set

* feature:merge:1506840174993

+++ cli/suman-daemon.sh
@@ -7,23 +7,25 @@ fi
 
 #set -e;
 
-# TODO: we could use either $(npm root -g)/suman-daemon or .suman/global/node_modules/suman-daemon
-
 echo "checking if existing process is listening on port"
-which_false=$(which false); # /bin/false or /usr/bin/false depending on system
+# which_false=$(which false); # /bin/false or /usr/bin/false depending on system
 
-pkill -f suman-daemon  # kill any existing suman-daemon process
 
-nc -zv localhost 9091  > /dev/null 2>&1
-nc_exit=$?
+if [[ "$1" != "--force" ]]; then
 
-echo "nc_exit => $nc_exit"
+    nc -zv localhost 9091  > /dev/null 2>&1
+    nc_exit=$?
 
-if [ ${nc_exit} -eq 0 ]; then
-    echo "a process is already listening on the default port"
-    echo "please choose another port with --port=x"
-    echo "suman-daemon may already be running - check with 'ps aux | grep suman-daemon'"
-    exit 1;
+    if [ ${nc_exit} -eq 0 ]; then
+        echo "a process is already listening on the default port"
+        echo "please choose another port with --port=x"
+        echo "suman-daemon may already be running - check with 'ps aux | grep suman-daemon'"
+        echo "otherwise, use '$ suman-daemon --force'."
+        exit 1;
+    fi
+
+else
+    pkill -f suman-daemon  # kill any existing suman-daemon process
 fi
 
 
@@ -31,20 +33,19 @@ fi
 # we have to use the global version,
 # because otherwise we would not know which suman installation to pre-load
 
-mkdir -p ~/.suman/global
-mkdir -p ~/.suman/logs
-
+mkdir -p "$HOME/.suman/global"
+mkdir -p "$HOME/.suman/logs"
 
 NPM_ROOT_GLOBAL="$(npm root -g)";
 
-export NODE_PATH=${NODE_PATH}:"~/.suman/global/node_modules"
-export PATH="~/.suman/global/node_modules/.bin":"${NPM_ROOT_GLOBAL}/suman-daemon/node_modules/.bin":${PATH}
+export NODE_PATH=${NODE_PATH}:"$HOME/.suman/global/node_modules"
+export PATH="$HOME/.suman/global/node_modules/.bin":"${NPM_ROOT_GLOBAL}/suman-daemon/node_modules/.bin":${PATH}
 export SUMAN_LIBRARY_ROOT_PATH="${NPM_ROOT_GLOBAL}/suman";
 
-WHICH_FOREVER=$(which forever);
+WHICH_FOREVER="$(which forever)";
 
 if [[ -z ${WHICH_FOREVER} ]]; then
- (cd ~/.suman/global && npm install forever)
+ (cd "$HOME/.suman/global" && npm install forever)
 fi
 
 if [[ -L "${NPM_ROOT_GLOBAL}/suman" || -d "${NPM_ROOT_GLOBAL}/suman" ]]; then
@@ -52,25 +53,26 @@ if [[ -L "${NPM_ROOT_GLOBAL}/suman" || -d "${NPM_ROOT_GLOBAL}/suman" ]]; then
 else
     # we need to install suman globally so that suman-daemon always pre-loads the same version of suman
     echo "suman is not installed globally, we will install suman globally now.";
-    npm install -g suman
+    npm install -g suman;
 fi
 
 daemon_log="$HOME/.suman/logs/suman-daemon.log";
 
-echo "begin of new daemon process" > ${daemon_log};
+echo "beginning of new daemon process" > ${daemon_log};
 
 if [[ -L "${NPM_ROOT_GLOBAL}/suman-daemon" || -d "${NPM_ROOT_GLOBAL}/suman-daemon" ]]; then
 
-    echo "found suman-daemon global"
-    node "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" > ${daemon_log} 2>&1 # &
-#      forever start "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" --workingDir $(pwd)
+    echo "found suman-daemon global installation."
+    echo "now starting suman-daemon..."
+    node "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" > "${daemon_log}" 2>&1 # &
+    # forever start "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" --workingDir $(pwd)
 
 else
 
    echo "installing suman-daemon globally, use --force-local to enforce local installations.";
    npm install -g suman-daemon &&
-   node "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" > ${daemon_log} 2>&1  #&
-#    forever start "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" --workingDir $(pwd)
+   node "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" > "${daemon_log}" 2>&1  #&
+   # forever start "${NPM_ROOT_GLOBAL}/suman-daemon/index.js" --workingDir $(pwd)
 
 fi
